Move legacy DV in and start to connect the pipes
Working through the hypermedia request spec trying to get different
parts of it to pass. Added a TODO for where I am now.

Also made a MemoryModel parent class for models which will have UUIDs
and perhaps pooling in the future.

Also made a MoveGenerator superclass for move generators. Might be
confusing vs MovesGenerator but should be fine for now.