Change routes to be uuid lookup only
Fixed the grape stuff enough to be able to join! Getting so close to
being playable again...

Also fixed a bug where the class-specific-scope wasn't working because
of how I was stubbing so I adjusted it slightly to make it easier to
stub out only the global stub. Turns out child classes inherit the stubs
of their parents!